requirements.txt
# Core API and Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
streamlit==1.28.1

# HTTP Client for Remote LLM API
httpx==0.25.2
requests==2.31.0

# Vector Database and Search
chromadb==0.4.18
llama-index==0.9.13

# Document Processing Libraries
python-docx==1.1.1
PyMuPDF==1.24.14
pdfminer.six==20231228
scikit-learn==1.5.2

# Environment Configuration
python-dotenv==1.0.0

# Data Processing (Lightweight)
pandas==2.1.3
numpy==1.25.2

# Document Processing
python-multipart==0.0.6
PyPDF2==3.0.1

# Logging and Validation
loguru==0.7.2
pydantic==2.11.7
pydantic-settings==2.9.1

# Authentication and Security
PyJWT==2.8.0
bcrypt==4.1.2

# CORS Support
fastapi-cors==0.0.6 